Do not pass $archiveName two times to OldLocalFile
Check the first dozen lines in ImportableUploadRevisionImporter::import().
In all cases the $archiveName is passed as a parameter to
OldLocalFile::newFromArchiveName(), and available via
OldLocalFile::getArchiveName().

I used https://codesearch.wmflabs.org/search/?q=uploadOld to make sure
this method is not called anywhere else. I consider it "package private"
and suggest to not apply the deprecation policy on this particular code.

Change-Id: Ibe9fb55d7e0d6e87698a56f7cfbf31d3d44ba70c