logging: Start using LinkTarget & UserIdentity in ManualLogEntry
This is a first, very small migration step. This is entirely
non-breaking, as all objects passed to these two functions already
matched the interfaces that are now required.

For the setTarget() method it's most obvious it really just needs a
LinkTarget value object, but not the entirety of the Title class
with all it's helper methods. Same for the performer. A ManualLogEntry
does not need to interact with the performer, just needs to know him.

This small change allows to slowly migrate away from code depending on
the User and Title classes.

Note this code currently does not have any test coverage, and I do not
plan to change this within this patch. I just don't have the time.

https://doc.wikimedia.org/cover/mediawiki-core/includes/logging/LogEntry.php.html

Change-Id: Id551f066ab02d5ca14346a39a55af4e7b6e4e042