Setup: Avoid using count() function in any kind of loop(s)
Using count() function in loops makes things very slow because
of function overheads and this function gets called everytime the
loop runs meaning the bigger the value of the variable in count(),
the slower the loop as its value always gets computed as the loop
runs.

In this case, the use of foreach(...){} is possible in order to
perform the computation making it pretty fast hence improving the
performance.

Change-Id: Ie21fbf8f6acf72373d1da75023725b4592c80386