Disallow user suppression in a partial block
Hide the "hide user" checkbox on Special:Block if the block
parameters specify a partial block. Return an error message
if a partial block with user suppression is specified via
the API. Also, make sure that an existing user suppression
is lifted if the user's block is reset to a partial block.

Bug: T210002
Change-Id: If42edfc85580d6bad6f4b397517e77e63deecc87