Correct namespace 'Template' for gom-deva
The existing name for the 'Template' namespace was incorrect. Hence,
replaced it with the correct word for 'Template' and 'Template_talk',
and added the existing ones as aliases.

Noticed that the existing file was missing a comma at the end of the
last item in the alias list.

Bug: T213717
Change-Id: Ibb8385cae3368e2c36006cf9993dec4ebbbc4491
(cherry picked from commit 38ba553407b02a8867eea7c1e3412f9f13d163f7)