Simplify boolean logic for Special:Block widgets
The boolean logic controlling the showing/hiding of widgets on
Special:Block could be expressed more simply. The logic is
becoming more complex due to partial blocks, so now is a good
time to rephrase it.

This removes safeguards against passing undefined to toggle,
so make sure none of the checks ever return undefined.

Change-Id: I1e0eb288db708be951fe5cf11f658725a38a7c4d