Datatable: Apply `padding` only on `th` & `td` elements
Applying it on `table` as well results in issues when wanting
to override `border-radius` together with `border-collapse`.
It's not needed in a well-formed table as every child element would
need to reside within a `th` or `td` anyways.
Also the only other content carrying element `caption` doesn't receive
table's `padding` by default.

Change-Id: I5e8c9d47fab806b56e6d03f1611ddbf6d4ea8924