build: Remove unused 'deepmerge' package
Was added for wdio-related code, but not used in the end.
The only reference to this module anywhere in Wikimedia Git is
mediawiki/extensions/CirrusSearch, which doesn't use it for
its regular or daily tests, but rather the integration test,
which already has its own copy of this module in package.json.

Change-Id: Ifdf4362077f4684a2866363e35d0aca2f89f48b5