tests: Remove use of $IP and MediaWikiTestCase in libs/composer tests
Reference the fixture files relatively instead of absolutely,
this is done for most other references to data files as well
I believe.

Use plain arrayEquals(), which seems to suffice here. In PHP 7 at
least, regular == doesn't require the declaration order of the
keys to be the same in order to evaluate to true.

Change-Id: Iddc874ec811f5c960e13d480d70bcb20334cfa1e