Document Uri.js constructor and methods throwing Errors
I, personally, think the fact an empty `new Uri()` call can fail with
an Error being thrown is a bug. But it seems others think this is a
feature (see Ib4dc568). Can we at least document it then? And then
somehow (not sure how) tell everybody who ever used this class to fix
their code?

Change-Id: I86ef1ccf90f9be9c61bc6a8312369ae09d622b0c