docs: Remove 'code-coverage' placeholder directory
In 8 years, this is the first time I see this directory exists
and that there is a Makefile that would write to it.

I don't know if anyone uses that, but PHPUnit auto-creates this
directory as needed. Hence, it need not exist ahead of time.

It seems most contributors generate code coverage by invoking
PHPUnit directory, which requires a path to be specified.

Per https://mediawiki.org/wiki/Manual:PHP_unit_testing/Code_coverage,
this tends to be docs/coverage, not docs/code-coverage.

Change the Makefile to write there as well, as better example,
and also add it to gitignore.

Change-Id: I0a4cf716ea9b7fae89c282945b160b0dc7b2d02f