watcheditem: Replace error message strings with constant string
As the messages are the same and in addition using the same class
DBReadOnlyError, make the error message a constant and pass it as
self::DB_READONLY_ERROR argument to this class. Changing the error
will apply to the class as intended rather than change it in each
single place that this string appears.

Change-Id: I7f978cc334020746b4c4405fa2a0a73219bdacac