FormatJson: Improve parse() error code handling and tests
Three of the errors are encode errors that won't be emitted when we're
trying to decode JSON, so we can ignore those lines of code.

JSON_ERROR_UTF16 is a new error code in PHP 7.0, so add that in.

Improve test coverage while we're at it. The UTF16 test case was
copied from php-src/ext/json/tests/bug62010.phpt.

Change-Id: I79aa0db3d967d512611f8521bb052af36c3cda8e