Fix read-only mode for LoadBalancerSingle
For master connections, LoadBalancer unconditionally replaces the
connection’s read-only mode with the load balancer’s, so if we don’t
include the read-only reason in the parameters we pass to the parent
constructor, the parameter we set on the connection has no effect.

I attempted to write tests for this, but eventually had to stop wasting
time on it, since I couldn’t get anywhere. Add a warning to the top of
the class to discourage others from making the same mistake.

Change-Id: I481553fac4f0f13a429ad4bbfdf15007c8299af9