RCFilters: Export config vars in the RL modules where possible
Right now there are only a few settings that don't depend on the request
context, but we can at least export those. The tag data in
particular can get pretty big.

Downgrade the type hint for ChangeTags::tagDescription() and
tagLongDescriptionMessage() from IContextSource to MessageLocalizer, so
we can pass a ResourceLoaderContext to these functions. (Since
IContextSource extends MessageLocalizer, this won't break any callers.)

truncateTagDescription() can't be downgraded to MessageLocalizer because
it needs a Language object, so hack up a fake RequestContext when
calling it. It would be nice if this weren't necessary; perhaps we can
move getLanguage() from IContextSource up into MessageLocalizer?

Bug: T201574
Change-Id: I9b66e35de826a07aa9551ba285e64e4852293229