ChangesList: Remove spaces from <td> in EnhancedChangesListGroup.mustache
The special page RecentChanges can included in wiki content via

  {{Special:RecentChanges}}

In this case the generated content is parsed by a wikitext parser. The
wikitext parser converts

  <td>
    <span>Foo</span>
  </td>

to the HTML

  <td><p><span>Foo</span></p></td>

But the wikitext

  <td><span>Foo</span></td>

is converted to the HTML

  <td><span>Foo</span></td>

The module 'jquery.makeCollapsible' supports custom collapsible toggle
buttons but only when they are a direct child of the <td> element. When
the module does not find a custom collapsible toggle button it adds the
default button [Expand].

Because of the additional <p> element when including the recent changes
the module 'jquery.makeCollapsible' adds a default collapsible toggle
button.

This change removes the spaces between <td> and <span> to prevent the
<p> elements when including the recent changes.

Bug: T215737
Change-Id: Ida46956fc7f9ab923c542751ac13411380c02ab0