Revert "RCFilters: Export config vars in the RL modules where possible"
This reverts commit bf4f9d0313495dd67a53b9925edf9a3cc6ef36fe.

Reason for revert: causes issues with OAuth extension

The OAuth extension uses a hook to override certain tag messages, and
that hook uses RequestContext::getMain()->msg(), which tries to load
the current user. That's not allowed in load.php requests, so we get a
"Sessions are disabled for this entry point" error.

Change-Id: If58b7edfca116cd7a11f38556f74bed82609d96a