jquery.makeCollapsible.styles: Avoid FOUC for collapse button
The expand button on a <div> element has already a pseudo-class with the
expand text as content. For the collapse button this pseudo-class is
missing.

This change adds the missing pseudo-class for the collapse button to
prevent a missing button while loading.

The missing collapse button while loading is visible on
Special:RecentChanges. The expand button is always there.

Change-Id: I7fa80238d54ac43a96bf1d70bd42a0cccf90a73a