EnhancedChangesList: Use HTML/CSS for collapsing
The collapsing on the enhanced changes list now works without
JavaScript.

Keyboard navigation is still supported but only the space and not the
return toggles the collapsing toggle button.

The <input type="checkbox"> element needs an unique identifier in the
`id` attribute and the same value must be in the `for` attribute of the
<label> element. A simple counter in the class EnhancedChangesList
starts twice from the beginning if the recent changes get included as:

 {{Special:RecentChanges|enhanced=1}}
 {{Special:RecentChanges|enhanced=1}}

Therefore a random value is used as checkbox identifier.

The module 'jquery.makeCollapsible' and its classes `mw-collapsible`,
`mw-collapsed` and `mw-collapsible-toggle-collapsed` are not needed
anymore.

The icons from module 'mediawiki.icon' are directly included because the
module has fixed selectors which do not fit here.

Bug: T172618
Change-Id: Iafd27e5d760b78ae386d833946005f86cee8dd64