specials: Fix fatal MWNamespace exception on Special:Contributions
Namespaces with negative values (valid like -1 and -2 or invalid like
less than -2), should not be processed as they do not have associated
namespaces. Rather than throwing an exception, nicely tell the user
the namespace is a negative value and doesn't have an associated
namespace to it.

Also, don't show any results, just load the form again so user can do
another request but show reason for this behavior atop of form.

Bug: T150324
Change-Id: I525d305a4dabb040110894d3230eeb5e04ff8336