pagers: Fix and add missing @param documentation
Sometimes the code expects a specific subclass as it calls a specific
method that only exists in this subclass. If this is not the case, all
the code needs to know is that it got some special page.

Some of these constructors should be changed to expect a IContextSource
instead of a special page. But this is for another patch. Here in this
patch I intentionally touch documentation only.

Change-Id: I33bae9641a8871f359edccacf88f09cfe8e83139