Check normalization rules of usernames during signup
The username may need to be adjusted due to technical restrictions
for page titles: the first letter is capitalized, underscores are
changed to spaces, and so on. Some of these tweaks can be unwanted
by some users: for example "Matma Rex" and "matmarex" are both
acceptable usernames for me, but "Matmarex" is not.

Display a warning if that happens (if the user has JavaScript
enabled). No action is needed from the user, other than reading
the message.

Bug: T63416
Change-Id: I120f17ca3801879ad650e5a81e7a4c270540cd4f