mw-ui-radio, mw-ui-checkbox: Don't reset every font style
The following existing style:

.client-js .mw-ui-radio:not( #noop ) {
 * {
	font: inherit;
  }
}

is a strong reset, preventing any font style customization such
as a bold label for .mw-ui-radio.

Change it only to font-size as required for the original reason
for override. Same for .mw-ui-checkbox

I am not sure whether this override is required now, but not going
in to that now.

Change-Id: I5d1815bcfa897bdb03c81be11f63e1acb9412c10